refactor: commented out some stuff which isn't being used yet (#283)
* refactor: commented out some stuff which isn't being used yet

* semver: upping the version

* feat: tests
