+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "katelibby",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "description": "An IRC and Slack Bot",
   "repository": {
     "type": "git",
